==> notebooks/03_load_to_database.ipynb <==
# import 
from sqlalchemy import create_engine 
import pandas as pd
import getpass
import pymysql

# configuration 
user= 'root'
password =getpass.getpass("Provide your MySql Password: ")
host = 'localhost'
database = 'olist_ecommerce'

# mysql connection
engine = create_engine(f'mysql+pymysql://{user}:{password}@{host}/{database}')

# load csv then push to mysql
(pd.read_csv("olist_fact_sales.csv",encoding='utf-8')).to_sql("fact_orders",con=engine,if_exists="replace",index=False)
(pd.read_csv("olist_RFM_analysis.csv",encoding='utf-8')).to_sql("rfm_scores",con=engine,if_exists="replace",index=False)
