==> notebooks/04_export_for_tableau.ipynb <==
# import 
from sqlalchemy import create_engine
import pandas as pd
import getpass
import pymysql

# configuration 
user= 'root'
password = getpass.getpass("MySQL Password: ")
host= 'localhost'
database = 'olist_ecommerce'

# connect to mysql
engine= create_engine(f'mysql+pymysql://{user}:{password}@{host}/{database}')

# load data

df= pd.read_sql("SELECT * FROM customer_profitability",con=engine)

# Profitability_segment
df["profitability_segment"] = df.apply(
    lambda x :
        "High Profit Customers" if x.total_profit > 500 and x.profit_margin>=0.30 else 
        "Low Margin Customers" if x.total_profit > 0 else "Loss Making Customers", axis=1
    


)

# Export for Tableau 
df.to_csv("customer_profitability.csv",index=False)
